pyproject.toml
[project]
name = "on-policy-quantization"
version = "0.1.0"
description = "On policy quantization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "datasets>=4.4.1",
    "jupyterlab>=4.5.1",
    "lm-eval>=0.4.0",
    "matplotlib",
    "numpy",
    "peft>=0.18.0",
    "pydantic>=2.12.5",
    "pydantic-config[toml]",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "trl>=0.26.2",
    "wandb>=0.23.1",
]

[project.optional-dependencies]
gpu = ["torchao", "fbgemm-gpu-genai>=1.4.2", "bitsandbytes>=0.49.0"]
quant = ["optimum>=1.23.0", "autoawq>=0.2.0"]

[[tool.uv.index]]
name = "pytorch-nightly"
url = "https://download.pytorch.org/whl/nightly/cu129"
explicit = true

[tool.uv.sources]
pydantic-config = { git = "https://github.com/tom-pollak/pydantic_config.git" }
torchao = { index = "pytorch-nightly" }
